# ENH: Improve fftn performance