--- a/scipy/fft/_pocketfft/tests/test_basic.py
+++ b/scipy/fft/_pocketfft/tests/test_basic.py
@@ -747,12 +747,12 @@ class TestFftn(object):
     def test_invalid_sizes(self):
         with assert_raises(ValueError,
                            match="invalid number of data points"
-                           r" \(\[1 0\]\) specified"):
+                           r" \(\[1, 0\]\) specified"):
             fftn([[]])
 
         with assert_raises(ValueError,
                            match="invalid number of data points"
-                           r" \(\[ 4 -3\]\) specified"):
+                           r" \(\[4, -3\]\) specified"):
             fftn([[1, 1], [2, 2]], (4, -3))
 
     def test_no_axes(self):
@@ -794,12 +794,12 @@ class TestIfftn(object):
     def test_invalid_sizes(self):
         with assert_raises(ValueError,
                            match="invalid number of data points"
-                           r" \(\[1 0\]\) specified"):
+                           r" \(\[1, 0\]\) specified"):
             ifftn([[]])
 
         with assert_raises(ValueError,
                            match="invalid number of data points"
-                           r" \(\[ 4 -3\]\) specified"):
+                           r" \(\[4, -3\]\) specified"):
             ifftn([[1, 1], [2, 2]], (4, -3))
 
     def test_no_axes(self):
@@ -839,12 +839,12 @@ class TestRfftn(object):
     def test_invalid_sizes(self, func):
         with assert_raises(ValueError,
                            match="invalid number of data points"
-                           r" \(\[1 0\]\) specified"):
+                           r" \(\[1, 0\]\) specified"):
             func([[]])
 
         with assert_raises(ValueError,
                            match="invalid number of data points"
-                           r" \(\[ 4 -3\]\) specified"):
+                           r" \(\[4, -3\]\) specified"):
             func([[1, 1], [2, 2]], (4, -3))
 
     @pytest.mark.parametrize('func', [rfftn, irfftn])