@peterbell10 2019-07-19
>To avoid duplication of the logic, can't you have `scipy.fftpack.helper._init_nd_shape_and_axes` just wrap this new version and cast the outputs to `ndarray`?

I'm trying to avoid circular dependecies between the two. However, I suppose if I just move the function into `fftpack` then we can have it both ways.