@larsoner 2019-07-19
> I'm trying to avoid circular dependecies between the two. However, I suppose if I just move the function into fftpack then we can have it both ways.

The nicer way to separate it I think would be to do a nested import in the `fftpack` code.

That way we stick as close as possible to the idea that "still maintained/improved code" lives in `scipy.fft` and legacy code lives in `scipy.fftpack`. Moreover, eventually hopefully someday soon importing `scipy.fft` won't require import anything from `scipy.fftpack`.