@peterbell10 2019-07-19
>The nicer way to separate it I think would be to do a nested import in the fftpack code.

By that do you mean just moving the imports inside functions? In that case I would say there's still a circular dependency; it just doesn't manifest itself as import issues.

>eventually hopefully someday soon importing `scipy.fft` won't require import anything from `scipy.fftpack`.

If we get dct/dst into `_pocketfft` (#10493) then `fftpack` can depend on `fft` without issue, so hopefully we can fix this very soon.